requirements.txt
-f /usr/share/pip-wheels
apiai==1.0.2
cffi==1.7.0
cryptography==1.4
Django==1.10
djangorestframework==3.4.4
google-api-python-client==1.5.2
httplib2==0.9.2
idna==2.1
numpy==1.11.1
oauth2client==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
pyowm==2.3.2
PySocks==1.5.7
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.0
requests-toolbelt==0.7.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
websocket-client==0.37.0
wit==4.0.0
